bigquery_loader_function/requirements.txt
google-cloud-bigquery
google-cloud-bigquery>=3.13.0